Move null saving specific tests into their own suite
Allows running non-null saving tests against 2.1
