+++ integration-tests/src/test/java/com/datastax/oss/driver/mapper/UpdateIT.java
@@ -28,12 +28,10 @@ import com.datastax.oss.driver.api.mapper.MapperException;
 import com.datastax.oss.driver.api.mapper.annotations.Dao;
 import com.datastax.oss.driver.api.mapper.annotations.DaoFactory;
 import com.datastax.oss.driver.api.mapper.annotations.DaoKeyspace;
-import com.datastax.oss.driver.api.mapper.annotations.DefaultNullSavingStrategy;
 import com.datastax.oss.driver.api.mapper.annotations.Mapper;
 import com.datastax.oss.driver.api.mapper.annotations.Select;
 import com.datastax.oss.driver.api.mapper.annotations.Update;
 import com.datastax.oss.driver.api.mapper.entity.saving.NullSavingStrategy;
-import com.datastax.oss.driver.api.testinfra.CassandraRequirement;
 import com.datastax.oss.driver.api.testinfra.ccm.CcmRule;
 import com.datastax.oss.driver.api.testinfra.session.SessionRule;
 import com.datastax.oss.driver.categories.ParallelizableTests;
@@ -49,7 +47,6 @@ import org.junit.rules.RuleChain;
 import org.junit.rules.TestRule;
 
 @Category(ParallelizableTests.class)
-@CassandraRequirement(min = "2.2", description = "support for unset values")
 public class UpdateIT extends InventoryITBase {
 
   private static CcmRule ccm = CcmRule.getInstance();
@@ -463,71 +460,6 @@ public class UpdateIT extends InventoryITBase {
     assertThat(dao.findById(FLAMETHROWER.getId()).getDimensions()).isNull();
   }
 
-  @Test
-  public void
-      should_update_entity_and_set_null_field_preferring_default_strategy_when_specific_not_set() {
-    // given
-    DefaultNullStrategyDao dao = inventoryMapper.defaultNullStrategyDao(sessionRule.keyspace());
-    assertThat(dao.findById(FLAMETHROWER.getId())).isNull();
-    dao.update(FLAMETHROWER);
-    assertThat(dao.findById(FLAMETHROWER.getId()).getDescription()).isNotNull();
-
-    // when
-    dao.update(new Product(FLAMETHROWER.getId(), null, FLAMETHROWER.getDimensions()));
-
-    // then
-    assertThat(dao.findById(FLAMETHROWER.getId()).getDescription()).isNull();
-  }
-
-  @Test
-  public void
-      should_update_entity_and_not_set_null_field_preferring_method_strategy_when_both_are_set() {
-    // given
-    DefaultNullStrategyDao dao = inventoryMapper.defaultNullStrategyDao(sessionRule.keyspace());
-    assertThat(dao.findById(FLAMETHROWER.getId())).isNull();
-    dao.update(FLAMETHROWER);
-    assertThat(dao.findById(FLAMETHROWER.getId()).getDescription()).isNotNull();
-
-    // when
-    dao.updateOverrideDefaultNullSavingStrategy(
-        new Product(FLAMETHROWER.getId(), null, FLAMETHROWER.getDimensions()));
-
-    // then
-    assertThat(dao.findById(FLAMETHROWER.getId()).getDescription()).isNotNull();
-  }
-
-  @Test
-  public void
-      should_update_entity_and_do_not_set_null_field_when_both_default_and_method_level_not_explicitly_set() {
-    // given
-    DoNotSetSavingStrategyDao dao = inventoryMapper.notSetNullStrategyDao(sessionRule.keyspace());
-    assertThat(dao.findById(FLAMETHROWER.getId())).isNull();
-    dao.update(FLAMETHROWER);
-    assertThat(dao.findById(FLAMETHROWER.getId()).getDescription()).isNotNull();
-
-    // when
-    dao.update(new Product(FLAMETHROWER.getId(), null, FLAMETHROWER.getDimensions()));
-
-    // then
-    assertThat(dao.findById(FLAMETHROWER.getId()).getDescription()).isNotNull();
-  }
-
-  @Test
-  public void should_update_entity_and_set_null_field_preferring_specific_over_default() {
-    // given
-    MethodOverrideNullSavingStrategyDao dao =
-        inventoryMapper.methodOverrideNullStrategyDao(sessionRule.keyspace());
-    assertThat(dao.findById(FLAMETHROWER.getId())).isNull();
-    dao.update(FLAMETHROWER);
-    assertThat(dao.findById(FLAMETHROWER.getId()).getDescription()).isNotNull();
-
-    // when
-    dao.update(new Product(FLAMETHROWER.getId(), null, FLAMETHROWER.getDimensions()));
-
-    // then
-    assertThat(dao.findById(FLAMETHROWER.getId()).getDescription()).isNull();
-  }
-
   @Mapper
   public interface InventoryMapper {
     @DaoFactory
@@ -538,16 +470,6 @@ public class UpdateIT extends InventoryITBase {
 
     @DaoFactory
     ProductWithoutIdDao productWithoutIdDao(@DaoKeyspace CqlIdentifier keyspace);
-
-    @DaoFactory
-    DefaultNullStrategyDao defaultNullStrategyDao(@DaoKeyspace CqlIdentifier keyspace);
-
-    @DaoFactory
-    DoNotSetSavingStrategyDao notSetNullStrategyDao(@DaoKeyspace CqlIdentifier keyspace);
-
-    @DaoFactory
-    MethodOverrideNullSavingStrategyDao methodOverrideNullStrategyDao(
-        @DaoKeyspace CqlIdentifier keyspace);
   }
 
   @Dao
@@ -612,39 +534,6 @@ public class UpdateIT extends InventoryITBase {
   public interface OnlyPKDao {
     @Update
     void update(OnlyPK onlyPK);
-  }
-
-  @Dao
-  @DefaultNullSavingStrategy(NullSavingStrategy.SET_TO_NULL)
-  public interface DefaultNullStrategyDao {
-
-    @Update
-    void update(Product product);
-
-    @Update(nullSavingStrategy = NullSavingStrategy.DO_NOT_SET)
-    void updateOverrideDefaultNullSavingStrategy(Product product);
-
-    @Select
-    Product findById(UUID productId);
-  }
-
-  @Dao
-  public interface DoNotSetSavingStrategyDao {
-    @Update
-    void update(Product product);
-
-    @Select
-    Product findById(UUID productId);
-  }
-
-  @Dao
-  @DefaultNullSavingStrategy(NullSavingStrategy.DO_NOT_SET)
-  public interface MethodOverrideNullSavingStrategyDao {
-    @Update(nullSavingStrategy = NullSavingStrategy.SET_TO_NULL)
-    void update(Product product);
-
-    @Select
-    Product findById(UUID productId);
   }
 
   @Dao
